Test: RankingRepositoryTest 통합테스트로 묶음
