--- src/test/groovy/com/stoury/repository/FeedRepositoryTest.groovy
@@ -1,103 +1,0 @@
-package com.stoury.repository
-
-import com.stoury.domain.Feed
-import com.stoury.domain.Member
-import org.springframework.beans.factory.annotation.Autowired
-import org.springframework.boot.test.autoconfigure.jdbc.AutoConfigureTestDatabase
-import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest
-import org.springframework.data.domain.PageRequest
-import spock.lang.Specification
-
-@DataJpaTest
-@AutoConfigureTestDatabase(replace = AutoConfigureTestDatabase.Replace.NONE)
-class FeedRepositoryTest extends Specification {
-    @Autowired
-    FeedRepository feedRepository
-    @Autowired
-    MemberRepository memberRepository
-
-    def member = new Member("aaa@dddd.com", "qwdqwdqwd", "username", null);
-
-    def setup() {
-        memberRepository.save(member)
-    }
-
-    def cleanup() {
-        feedRepository.deleteAll()
-        memberRepository.deleteAll()
-    }
-
-    def "해외에서 10개 인기 여행장소"() {
-        given:
-        (0..<3).each { i ->
-            def feed = new Feed(member, "feed#" + i + 8, 11.11, 11.11, Collections.emptyList(), "city", "country")
-            feed.country = "France"
-            feed.city = "Paris"
-            feedRepository.save(feed)
-        }
-        (0..<4).each { i ->
-            def feed = new Feed(member, "feed#" + i, 11.11, 11.11, Collections.emptyList(), "city", "country")
-            feed.country = "United States"
-            feed.city = "NY"
-            feedRepository.save(feed)
-        }
-        (0..<1).each { i ->
-            def feed = new Feed(member, "feed#" + i + 12, 11.11, 11.11, Collections.emptyList(), "city", "country")
-            feed.country = "Vietnam"
-            feed.city = "hanoi"
-            feedRepository.save(feed)
-        }
-        (0..<2).each { i ->
-            def feed = new Feed(member, "feed#" + i + 12, 11.11, 11.11, Collections.emptyList(), "city", "country")
-            feed.country = "Australia"
-            feed.city = "Sydney"
-            feedRepository.save(feed)
-        }
-        def page = PageRequest.of(0, 10)
-
-        when:
-        def feeds = feedRepository.findTop10CountriesNotKorea(page)
-        then:
-        feeds.get(0) == "United States"
-        feeds.get(1) == "France"
-        feeds.get(2) == "Australia"
-        feeds.get(3) == "Vietnam"
-    }
-
-    def "국내에서 10개 인기 여행장소"() {
-        given:
-        (0..<3).each { i ->
-            def feed = new Feed(member, "feed#" + i + 8, 11.11, 11.11, Collections.emptyList(), "city", "country")
-            feed.country = "South Korea"
-            feed.city = "Seoul"
-            feedRepository.save(feed)
-        }
-        (0..<4).each { i ->
-            def feed = new Feed(member, "feed#" + i, 11.11, 11.11, Collections.emptyList(), "city", "country")
-            feed.country = "South Korea"
-            feed.city = "Busan"
-            feedRepository.save(feed)
-        }
-        (0..<1).each { i ->
-            def feed = new Feed(member, "feed#" + i + 12, 11.11, 11.11, Collections.emptyList(), "city", "country")
-            feed.country = "South Korea"
-            feed.city = "Hwacheon"
-            feedRepository.save(feed)
-        }
-        (0..<2).each { i ->
-            def feed = new Feed(member, "feed#" + i + 12, 11.11, 11.11, Collections.emptyList(), "city", "country")
-            feed.country = "South Korea"
-            feed.city = "Daejeon"
-            feedRepository.save(feed)
-        }
-        def page = PageRequest.of(0, 10)
-
-        when:
-        def feeds = feedRepository.findTop10CitiesInKorea(page)
-        then:
-        feeds.get(0) == "Busan"
-        feeds.get(1) == "Seoul"
-        feeds.get(2) == "Daejeon"
-        feeds.get(3) == "Hwacheon"
-    }
-}
